feat!: tighten types and fix API surface across SDK
Change all As*() union accessors from (T, bool) to (T, error) so callers
can distinguish "wrong variant" (nil, nil) from "corrupt data" (nil, err).
Add Raw() methods to all union types for inspecting unknown variants.

Additional fixes:
- float64 → int64/int for token counts, line numbers, additions/deletions
- interface{} → json.RawMessage for opaque Parameters/Data fields
- Streaming errors now return *APIError instead of fmt.Errorf
- SessionNewParams → SessionCreateParams
- Remove duplicate App.Providers (use Config.Providers)
- Remove stale Stainless generation comments

BREAKING CHANGE: As*() methods now return (T, error) instead of (T, bool).
SessionNewParams renamed to SessionCreateParams. App.Providers removed.
Multiple float64 fields changed to int64/int.

+++ file.go
@@ -1,5 +1,3 @@
-// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
-
 package opencode
 
 import (
@@ -143,10 +141,10 @@ type FileReadResponsePatch struct {
 
 type FileReadResponsePatchHunk struct {
 	Lines    []string `json:"lines"`
-	NewLines float64  `json:"newLines"`
-	NewStart float64  `json:"newStart"`
-	OldLines float64  `json:"oldLines"`
-	OldStart float64  `json:"oldStart"`
+	NewLines int64    `json:"newLines"`
+	NewStart int64    `json:"newStart"`
+	OldLines int64    `json:"oldLines"`
+	OldStart int64    `json:"oldStart"`
 }
 
 type FileListParams struct {
